partea de server 	-> in caz de eroare le login 	-> la retragere din licitatie 	-> administrare 	-> corectat formule complexe

+++ trunk/Atlantida/Server/Elemente_atomice.java
@@ -12,7 +12,7 @@ public class Elemente_atomice extends BazaDeDate
 		super(numeBazaDate);
 	}
 	
-	public int inserareInregistrare(String numeElement, String propSpeciala, String formulaCalcul, int valMaxima, int pretElement)
+	public int inserareInregistrare(String numeElement, String propSpeciala, String formulaCalcul, float valMaxima, float pretElement)
 	{	
 		Statement decl = null;
 		ResultSet rezultat = null;
